+++ lib/fortnite.dart
@@ -5,6 +5,8 @@ export "resources/mcp_operations.dart";
 export "resources/auth_clients.dart";
 export "resources/account_resources.dart" show accountResources;
 export "resources/homebase_rating_keys.dart";
+export "resources/base_item_rating.dart";
+export "resources/survivor_item_rating.dart";
 
 export "src/structures/client_options.dart";
 export "src/structures/device_auth.dart";
@@ -25,6 +27,9 @@ export "src/structures/player.dart";
 export "src/structures/banner_quest.dart";
 export "src/structures/survivor_squad_preset.dart";
 export "src/structures/claim_daily_response.dart";
+export "src/structures/stw_ingredient.dart";
+export "src/structures/stw_world_item.dart";
+export "src/structures/stw_ammo.dart";
 
 export "src/util/utils.dart";
 export "src/util/curve_table.dart";
@@ -36,5 +41,8 @@ export "src/client/mcp_profile.dart";
 export "src/client/athena_profile.dart";
 export "src/client/common_core_profile.dart";
 export "src/client/campaign_profile.dart";
+export "src/client/inventory_profile.dart";
+export "src/client/backpack_profile.dart";
+export "src/client/storage_profile.dart";
 
 export "package:logging/logging.dart" show Level;